refactor(professionals): move resume parsing route to live under professionals
