--- src/resumes/resumes.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ResumesService } from './resumes.service';
-import { ResumesController } from './resumes.controller';
-
-@Module({
-  controllers: [ResumesController],
-  providers: [ResumesService],
-})
-export class ResumesModule {}
