--- src/resumes/resumes.service.ts
@@ -1,10 +1,0 @@
-import { parseResumeFromPdf } from '@/shared/resume-parser';
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class ResumesService {
-  async parseResume(fileBuffer: Buffer) {
-    const parsedResume = await parseResumeFromPdf(fileBuffer);
-    return parsedResume;
-  }
-}
